Issue #1475 Initial work to allow download of the OCP binary
This uses --openshift-version to allow download of the OCP binary
from mirror.openshift.org or developers.redhat.com.

This is WIP and provides only initial functionality; it does not
handle the compile process by tags, it does not offer version
checking (as OCP is not semver compatible), and still contains
code duplication with the GitHub download implementation.

Signed-off-by: Gerard Braad <me@gbraad.nl>